Add dist tags
When building the same package for multiple dists, it's necessary
to have different versions if the packages are going to live in
the same repo / share a cache (debs was pretty much made for that
case). So by default, append ~{RELEASE} to a package version so
that packages don't step on each other.